Isolate keepDataRangeTest
