+++ src/functions/searches/tests/index.ts
@@ -1,0 +1,1 @@
+export { keepDataRangeTest } from './keep-data-range-test.spec';
